+++ src/console/controllers/DefaultController.php
@@ -39,11 +39,10 @@ class DefaultController extends Controller
 		$schedule = new Schedule();
 
 		// Clear expired bookings
-		// FIXME: See BookingsService.php:128
-//		$schedule->queue(function () {
-//			echo 'Clearing Expired' . PHP_EOL;
-//			Bookings::getInstance()->bookings->clearExpiredBookings();
-//		})->everyMinute();
+		$schedule->queue(function () {
+			echo 'Clearing Expired' . PHP_EOL;
+			Bookings::getInstance()->bookings->clearExpiredBookings();
+		})->everyMinute();
 
 		// Update next available cache
 		$schedule->queue(function () {