+++ CHANGELOG.md
@@ -9,7 +9,7 @@ interval by your duration.
 ### Added
 - Added "By Day" field to calendar UI.
 - Added support for updating multiple booked tickets custom fields in a single
-request. 
+request.
 
 ### Changed
 - If event multiplier is 0, availability will show the capacity.
@@ -20,6 +20,7 @@ request.
 - Fixed day view showing duration end minutes incorrectly above a certain time [#21]
 - Fixed minutes that end on an hour from extending an extra hour [#22]
 - Fixed issues with availability checking when using a multiplier of 0.
+- Fixed booking expire command timing out.
 
 [#21]: https://github.com/ethercreative/bookings/issues/21
 [#22]: https://github.com/ethercreative/bookings/issues/22